requirements.txt
aioredis==1.3.1
appdirs==1.4.3
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.7.1
Automat==20.2.0
backports.entry-points-selectable==1.1.0
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.1
channels==2.4.0
channels-redis==3.0.1
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
contextlib2==0.6.0
cryptography==3.0
daphne==2.5.0
distlib==0.3.3
distro==1.4.0
Django==3.0.8
django-cors-headers==3.4.0
django-eventstream==3.1.0
django-filter==2.3.0
django-grip==3.0.0
djangorestframework==3.11.0
filelock==3.0.12
gripcontrol==4.0.0
hiredis==1.1.0
html5lib==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
ipaddr==2.2.0
lockfile==0.12.2
msgpack==1.0.2
packaging==20.3
pep517==0.8.2
pipenv==2020.8.13
platformdirs==2.4.0
progress==1.5
psycopg2-binary==2.8.5
pubcontrol==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
retrying==1.3.3
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.8
virtualenv==20.0.31
virtualenv-clone==0.5.4
webencodings==0.5.1
Werkzeug==0.16.1
zope.interface==5.1.0
